+++ core/src/main/kotlin/com/bumble/appyx/core/integrationpoint/NodeActivity.kt
@@ -23,7 +23,7 @@ abstract class NodeActivity : AppCompatActivity() {
         protected set
 
     protected open fun createIntegrationPoint(savedInstanceState: Bundle?): ActivityIntegrationPoint =
-        ActivityIntegrationPoint(
+        ActivityIntegrationPoint.createIntegrationPoint(
             activity = this,
             savedInstanceState = savedInstanceState
         )