+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Pending changes
 
 - [#119](https://github.com/bumble-tech/appyx/pull/119) - **Fixed**: Lifecycle observers are invoked in incorrect order (child before parent)
+- [#62](https://github.com/bumble-tech/appyx/pull/62) - **Fixed**: Node is marked with stable annotation making some of the composable functions skippable 
 
 ## 1.0-alpha06 – 26 Aug 2022